added authentication

+++ src/components/ApplicationViews.js
@@ -66,14 +66,14 @@ export default class ApplicationViews extends Component {
       .then(r => newState.friendsEvents = r)
       .then(() => this.setState(newState))
   }
-addTask = task => TaskManager.post(task).then(() => this.loadAllData(sessionStorage.getItem("userID")))
+  addTask = task => TaskManager.post(task).then(() => this.loadAllData(sessionStorage.getItem("userID")))
 
-onLogin = () => {
-  this.setState({
-    userId: sessionStorage.getItem("userID")
-  })
-  this.loadAllData(this.state.userId)
-}
+  onLogin = () => {
+    this.setState({
+      userId: sessionStorage.getItem("userID")
+    })
+    this.loadAllData(this.state.userId)
+  }
 
   isAuthenticated = () => sessionStorage.getItem("userID") !== null
 
@@ -91,7 +91,7 @@ onLogin = () => {
     messageData.update(editedMessage)
       .then(() => this.loadAllData())
   }
-  
+
   createEvent = (newEvent) => {
     return ResourceManager.postEntry(newEvent, "events")
       .then(() => ResourceManager.getAll("events", sessionStorage.getItem("userID")))
@@ -103,18 +103,18 @@ onLogin = () => {
   }
   updateEvent = (eventToUpdate) => {
     return ResourceManager.updateEntry(eventToUpdate, "events")
-    .then(() => ResourceManager.getAll("events", sessionStorage.getItem("userID")))
-    .then(events => {
-      this.setState({
-        events: events
+      .then(() => ResourceManager.getAll("events", sessionStorage.getItem("userID")))
+      .then(events => {
+        this.setState({
+          events: events
+        })
       })
-    })
   }
 
-  
+
   //function is called in ArticleAddNewForm. it performs a post request, then gets all updated data and setState to re render Articles with updated
-addItem = (path, object, currentUserId) => ResourceManager.postItem(path, object)
-.then(() => this.loadAllData(currentUserId))
+  addItem = (path, object, currentUserId) => ResourceManager.postItem(path, object)
+    .then(() => this.loadAllData(currentUserId))
 
 
   // The addFriend function is passed down to the userSearch component in the friends directory. It makes sure that the user isn't trying to add him/herself as a friend. Then it will ask if the user is sure he/she wants to add this user as a friend. If so, we will create the newFriend object. The userId is the id of the user passed in as an argument. The currentUserId is grabbed from sessionStorage. Then we make a POST to the friends collection of our database.JSON. 
@@ -122,26 +122,26 @@ addItem = (path, object, currentUserId) => ResourceManager.postItem(path, object
   // Next we get the updated list of the user's friends by doing a GET call. Then we take the promise value and update the local state with the response. This state will be used as a prop. If the username isn't found, it will alert the user.
 
   addFriend = (user) => {
-    if(user.userName){
-      if(user.id === Number(sessionStorage.getItem("userID"))){
+    if (user.userName) {
+      if (user.id === Number(sessionStorage.getItem("userID"))) {
         window.alert("You can't add yourself as a friend.")
       } else {
-          if(window.confirm(`Would you like to add ${user.userName} as a friend?`)){
-              const newFriend = {
-                userId: user.id,
-                currentUserId: Number(sessionStorage.getItem("userID"))
-              }
+        if (window.confirm(`Would you like to add ${user.userName} as a friend?`)) {
+          const newFriend = {
+            userId: user.id,
+            currentUserId: Number(sessionStorage.getItem("userID"))
+          }
 
-              ResourceManager.postItem("friends", newFriend)
+          ResourceManager.postItem("friends", newFriend)
             .then(() => ResourceManager.getFriendsUserId(Number(sessionStorage.getItem("userID"))))
             .then(friends => this.setState({
-                friends: friends
+              friends: friends
             }))
-            } else {
-                window.alert("Username not found")
-              }
-          }
+        } else {
+          window.alert("Username not found")
+        }
       }
+    }
   }
 
   // The deleteFriend function is passed down to the Friends Component. This handles the functionality of making a DELETE call, and then loading all the updated data from the database with loadAllData. Then updating local state with the promise value.
@@ -156,30 +156,30 @@ addItem = (path, object, currentUserId) => ResourceManager.postItem(path, object
 
   registerUser = (userToRegister) => {
     return ResourceManager.postEntry(userToRegister, "users")
-    .then(() => ResourceManager.getAllUsers())
-    .then(users => this.setState({
-      users: users
-    }))
+      .then(() => ResourceManager.getAllUsers())
+      .then(users => this.setState({
+        users: users
+      }))
   }
 
-//function is called when delete button is click, performs delete method, then re loads data with new state
-deleteItem = (path, id) => ResourceManager.deleteItem(path, id)
-.then(() => ResourceManager.getSortedArticles(sessionStorage.getItem("userID")))
-.then(r => {
-  this.setState({
-    [path]: r
-  })
-})
+  //function is called when delete button is click, performs delete method, then re loads data with new state
+  deleteItem = (path, id) => ResourceManager.deleteItem(path, id)
+    .then(() => ResourceManager.getSortedArticles(sessionStorage.getItem("userID")))
+    .then(r => {
+      this.setState({
+        [path]: r
+      })
+    })
 
-//function is called when edit form is saved. performs PUT method and re loads data with new state
-updateItem = (path, object) => ResourceManager.putItem(path, object)
-.then(() => ResourceManager.getSortedArticles(sessionStorage.getItem("userID")))
-.then(r => {
-  this.setState({
-    [path]: r
-  })
-})
-  
+  //function is called when edit form is saved. performs PUT method and re loads data with new state
+  updateItem = (path, object) => ResourceManager.putItem(path, object)
+    .then(() => ResourceManager.getSortedArticles(sessionStorage.getItem("userID")))
+    .then(r => {
+      this.setState({
+        [path]: r
+      })
+    })
+
   render() {
     return (
       <React.Fragment>
@@ -188,26 +188,29 @@ updateItem = (path, object) => ResourceManager.putItem(path, object)
           exact path="/login" render={props => {
             return <Login users={this.state.users}
               onLogin={this.onLogin} {...props} />
-            
+
           }}
         />
         <Route
           exact path="/register" render={props => {
             return <Register users={this.state.users}
               registerUser={this.registerUser} onLogin={this.onLogin} {...props} />
-            
+
           }}
         />
 
-        <Route
-          exact path="/articles" render={props => {
+        <Route exact path="/articles" render={props => {
+          if (this.isAuthenticated()) {
             return <Articles articles={this.state.articles} friendsArticles={this.state.friendsArticles} {...props} addItem={this.addItem} deleteItem={this.deleteItem} users={this.state.users} />
-            // Remove null and return the component which will show news articles
-          }}
-        />
+          } else {
+            return <Redirect to="/login" />
+          }
+        }} />
+
         <Route exact path="/articles/new" render={(props) => {
           return <ArticleAddNewForm addItem={this.addItem} {...props} />
         }} />
+
         <Route path="/articles/edit/:articleId(\d+)" render={(props) => {
           return <ArticleEditForm updateItem={this.updateItem} {...props} />
         }} />
@@ -216,7 +219,7 @@ updateItem = (path, object) => ResourceManager.putItem(path, object)
           path="/friends" render={props => {
             if (this.isAuthenticated()) {
               return <FriendsList {...props} friends={this.state.friends} addFriend={this.addFriend}
-              deleteFriend={this.deleteFriend}/>
+                deleteFriend={this.deleteFriend} />
             } else {
               return <Redirect to="/login" />
             }
@@ -245,7 +248,7 @@ updateItem = (path, object) => ResourceManager.putItem(path, object)
 
         <Route
           exact path="/events" render={props => {
-            return <EventList {...props} users = {this.state.users} friendsEvents = {this.state.friendsEvents} events={this.state.events} />
+            return <EventList {...props} users={this.state.users} friendsEvents={this.state.friendsEvents} events={this.state.events} />
           }}
         />
         <Route
@@ -268,10 +271,10 @@ updateItem = (path, object) => ResourceManager.putItem(path, object)
           }}
         />
 
-        <Route path="/tasks/new" render={(props)=> {
+        <Route path="/tasks/new" render={(props) => {
           return <TaskForm {...props}
             addTask={this.addTask}
-            />
+          />
         }} />
       </React.Fragment>
     );
